% Generated by roxygen2 (4.0.1): do not edit by hand
\name{startWebServer}
\alias{startWebServer}
\title{Start a R web server}
\usage{
startWebServer(path)
}
\arguments{
\item{path}{path to the application folder.}
}
\description{
\code{startWebServer} starts a web server at the desired location. Before using
you should have created a directory for the application with
\code{\link{webApp.skeleton}}.
}

