% Generated by roxygen2 (4.0.1): do not edit by hand
\name{webApp.skeleton}
\alias{webApp.skeleton}
\title{Create a skeleton for a new web application}
\usage{
webApp.skeleton(name = "myRWebApp", path = ".", port = 8080,
  force = FALSE)
}
\arguments{
\item{name}{the name of the directory to create}

\item{path}{path where to create the directory}

\item{port}{port that your application will be listening for}

\item{force}{if TRUE and the directory already exists it will be overwritten}
}
\description{
\code{webApp.skeleton} creates a foldeer with all configuration files and
directories that will be needed by the R web server.
}
\details{
In the application directory, \code{webApp.skeleton} will create the following
files and folders :
\itemize{
\item rserve.conf: Rserve configuration file. you should not edit it.
\item onStartWebApp.R: script that is executed when server starts. You should
not edit it.
\item packages: list of the libraries to load.
\item routes: Routes for your application
\item lib: All R files in this folder will be sourced when the application
starts. Useful to define functions or constants.
\item data: all data files in this folder will be loaded when the application
starts. The data files have to be created with the "save" function.
\item R: This folder contains the R files that will be executed when the
application receives a query. Each script have to define at least a function
called "run". When the server receives a query like
"http://myserver:8080/myRscript?x=1y=2", it will source the script "myScript.R"
and execute "\code{run(x="1", y="2")}".
}
}

