% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WebResult}
\alias{WebResult}
\title{Result object of a FastRWeb script}
\usage{
WebResult(cmd = "html", payload = "",
  content.type = "text/html; charset=utf-8", headers = character(0))
}
\arguments{
\item{cmd}{string, command passed back to the FastRWeb
interface. Currently supported commands are \code{"html"},
\code{"file"}, \code{"tmpfile"} and \code{"raw"}. See details below.}

\item{payload}{string, the body (contents) that will be sent back or file name,
depending on the command}

\item{content.type}{MIME content type specification as it will be returned to the browser}

\item{headers}{string vector, optional additional headers to be sent
to the browser. Must not contain CR or LF!}

\item{x}{object to convert into \code{WebResult}}

\item{\dots}{additional arguments passed to the method}
}
\description{
\code{WebResult} is the class of the object that will be returned
from the \code{run} function of a FastRWeb script back to the browser.
}
\details{
Using a separate class allows automatic conversion of other objects
into the necessary representation - all that is needed is a
\code{as.WebResult} method for that particular object.

\code{WebResult} function can be used to create such objects directly.

\code{as.WebResult} coerces an object into a \code{WebResult}, it is a
generic. This allows methods to be defined for \code{as.WebResult}
which act as convertors transforming R objects into web results.

glop glop
}

